Pessimistic semver deps; drop OJ
After not having touched this project in a while, I come back to some
convoluted issue with either Oj or MultiJson's use of it. I gave it two
minutes but wasn't willing to invest further effort; I'll just use the
json gem and be done with it.

I'm not hip to all the different issues across ruby implementations, so
next I'll turn on travis and see what happens.

For future reference:

    ArgumentError:
      :indent must be a Fixnum.
    # <snip>/multi_json-1.10.1/lib/multi_json/adapters/oj.rb:21:in `dump'